+++ packages/css-throttle/src/index.ts
@@ -1,0 +1,7 @@
+function useCssThrottle() {
+	return {
+		//
+	}
+}
+
+export default useCssThrottle