+++ CHANGELOG.md
@@ -2,5 +2,5 @@
 
 ## 2023.06.06 v1.0.0
 
-1. add `use-downloads`
+1. add `use-downloads` `use-textarea` `use-speak` `use-css-throttle`
 2. api docs: [README.md](./README.md)